+++ src/textual/dom.py
@@ -1,6 +1,7 @@
 """
-A DOMNode is a base class for any object within the Textual Document Object Model,
+The module contains `DOMNode`, the base class for any object within the Textual Document Object Model,
 which includes all Widgets, Screens, and Apps.
+
 """
 
 from __future__ import annotations
@@ -31,7 +32,6 @@ from rich.tree import Tree
 from ._context import NoActiveAppError, active_message_pump
 from ._node_list import NodeList
 from ._types import WatchCallbackType
-from ._worker_manager import WorkerManager
 from .binding import Binding, BindingType, _Bindings
 from .color import BLACK, WHITE, Color
 from .css._error_tools import friendly_list
@@ -41,9 +41,10 @@ from .css.parse import parse_declarations
 from .css.styles import RenderStyles, Styles
 from .css.tokenize import IDENTIFIER
 from .message_pump import MessagePump
-from .reactive import Reactive, ReactiveError, _watch
+from .reactive import Reactive, ReactiveError, _Mutated, _watch
 from .timer import Timer
 from .walk import walk_breadth_first, walk_depth_first
+from .worker_manager import WorkerManager
 
 if TYPE_CHECKING:
     from typing_extensions import Self, TypeAlias
@@ -252,6 +253,10 @@ class DOMNode(MessagePump):
         this method after your reactive is updated. This will ensure that all the reactive _superpowers_
         work.
 
+        !!! note
+
+            This method will cause watchers to be called, even if the value hasn't changed.
+
         Args:
             reactive: A reactive property (use the class scope syntax, i.e. `MyClass.my_reactive`).
         """
@@ -277,6 +282,7 @@ class DOMNode(MessagePump):
                 yield WorldClock("Europe/Paris").data_bind(WorldClockApp.time)
                 yield WorldClock("Asia/Tokyo").data_bind(WorldClockApp.time)
             ```
+
 
         Raises:
             ReactiveError: If the data wasn't bound.
@@ -333,7 +339,8 @@ class DOMNode(MessagePump):
                     """Set bound data."""
                     _rich_traceback_omit = True
                     Reactive._initialize_object(self)
-                    setattr(self, variable_name, value)
+                    # Wrap the value in `_Mutated` so the setter knows to invoke watchers etc.
+                    setattr(self, variable_name, _Mutated(value))
 
                 return setter
 
@@ -513,7 +520,7 @@ class DOMNode(MessagePump):
         """Get a "component" styles object (must be defined in COMPONENT_CLASSES classvar).
 
         Args:
-            name: Name of the component.
+            names: Names of the components.
 
         Raises:
             KeyError: If the component class doesn't exist.
@@ -987,16 +994,17 @@ class DOMNode(MessagePump):
 
         for node in reversed(self.ancestors_with_self):
             styles = node.styles
+            has_rule = styles.has_rule
             opacity *= styles.opacity
-            if styles.has_rule("background"):
+            if has_rule("background"):
                 text_background = background + styles.background
                 background += styles.background.multiply_alpha(opacity)
             else:
                 text_background = background
-            if styles.has_rule("color"):
+            if has_rule("color"):
                 color = styles.color
             style += styles.text_style
-            if styles.has_rule("auto_color") and styles.auto_color:
+            if has_rule("auto_color") and styles.auto_color:
                 color = text_background.get_contrast_text(color.a)
 
         style += Style.from_color(
@@ -1536,7 +1544,7 @@ class DOMNode(MessagePump):
 
         Args:
             action: The name of an action.
-            action_parameters: A tuple of any action parameters.
+            parameters: A tuple of any action parameters.
 
         Returns:
             `True` if the action is enabled+visible,