+++ src/textual/strip.py
@@ -1,5 +1,7 @@
 """
-A Strip contains the result of rendering a widget.
+This module contains the `Strip` class and related objects.
+
+A `Strip` contains the result of rendering a widget.
 See [line API](/guide/widgets#line-api) for how to use Strips.
 """