+++ tests/test_geometry.py
@@ -62,6 +62,14 @@ def test_clamp():
     assert clamp(10, 0, 10) == 10
     assert clamp(5, 10, 0) == 5
 
+    # range in reverse order
+    assert clamp(5, 10, 0) == 5
+    assert clamp(-1, 10, 0) == 0
+    assert clamp(11, 10, 0) == 10
+    assert clamp(0, 10, 0) == 0
+    assert clamp(10, 10, 0) == 10
+    assert clamp(5, 0, 10) == 5
+
 
 def test_offset_bool():
     assert Offset(1, 0)